requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
pytorch-lightning>=2.0.0

# Data and utilities
pandas>=2.0.0
pyyaml>=6.0
hydra-core>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Testing and quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
scipy>=1.11.0
